[ADD] Create script to shut Franka down
- [REF] Refactors scripts to have one base class
